Add DSignal 'Bundle' and 'delay' variant.
'delay' copies the 'Signal' variant, e.g. delays a 'DSignal' and values
after reset are undefined, in contrast to 'delayed' where values are set
to an explicit value.

'Bundle' follows 'Signal' variant, and is useful to avoid explicit
'Signal' to 'DSignal' conversions that are otherwise necessary.

Also includes 'pipelineTreeFold' which provides a tree structured fold
that inserts a flip-flop after every fold operation.